Add boot scene
