+++ package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "clean-webpack-plugin": "^3.0.0",
     "html-webpack-plugin": "^4.3.0",
-    "mini-css-extract-plugin": "^0.9.0"
+    "mini-css-extract-plugin": "^0.9.0",
+    "phaser": "^3.20.1"
   }
 }
