+++ src/index.js
@@ -1,7 +1,22 @@
 import './styles/style.css';
 
-import config from './config/config'
+import config from './Config/config';
+import GameScene from './Scenes/gameScene';
+import BootScene from './Scenes/BootScene';
+import PreloaderScene from './Scenes/PreloaderScene';
+import OptionsScene from './Scenes/Options';
+import CreditsScene from './Scenes/Credits';
 
-config
+class Game extends Phaser.Game {
+  constructor() {
+    super(config);
+    this.scene.add('Boot', BootScene);
+    this.scene.add('Preloader', PreloaderScene);
+    this.scene.add('Options', OptionsScene);
+    this.scene.add('Credits', CreditsScene);
+    this.scene.add('Game', GameScene);
+    this.scene.start('Game');
+  }
+}
 
-export default new Phaser.Game(config)
+window.game = new Game();
